Drop shelly dependency.
This commit drops the dependency of antigen-hs on shelly, allowing a much
simpler installation story. The package will now build with just the standard
GHC packages + text, and text is probably already installed globally because
cabal-install needs it.

I also changed the behavior of Development RepoStorage. I renamed it to Local,
and the behavior is basically: things inside a Local RepoStorage will be
sourced as-is, without copying over to ~/.antigen-hs. You still have to run
antigen-hs-compile to start sourcing anything new that was added to a Local
RepoStorage.